requirements.txt
colorlog==6.8.0
matplotlib==3.8.1
ndicts==0.3.0
numpy==1.26.4
openai==1.33.0
pandas==2.0.3
python-dotenv==1.0.1
scikit_learn==1.3.0
sentence_transformers==2.2.2
tqdm==4.66.1
tyro==0.6.4
